Move non-jinja helpers in editors to editors/utils.py
