+++ src/olympia/editors/management/commands/auto_approve.py
@@ -10,10 +10,10 @@ import waffle
 
 import olympia.core.logger
 from olympia import amo
-from olympia.editors.templatetags.jinja_helpers import ReviewHelper
 from olympia.editors.models import (
     AutoApprovalNotEnoughFilesError, AutoApprovalNoValidationResultError,
     AutoApprovalSummary, clear_reviewing_cache, set_reviewing_cache)
+from olympia.editors.utils import ReviewHelper
 from olympia.files.utils import atomic_lock
 from olympia.versions.models import Version
 from olympia.zadmin.models import get_config
